+++ CHANGELOG.md
@@ -2348,3 +2348,4 @@
 - [2026-06-26 09:05] perf(i18n): optimize memory consumption in i18n list viewport
 - [2026-06-26 09:28] feat(state): add async state stream provider
 - [2026-06-26 09:33] feat(auth): add async auth stream provider
+- [2026-06-26 09:43] test(auth): add unit tests for auth business logic